Fixed incorrect error handling in LJUSB_OpenAllDevices
The variable devh was initialized to NULL and then never changed. It was used in 2 calls to libusb_close(devh) but was always NULL when passed.  In fact, both calls to libusb_close() were using the wrong parameter, they should be libusb_close(handle).

Relatedly, the decrement of ljFoundCount in the error path just seems wrong, and would result in the function returning the wrong count of devices actually opened.