Code Review (#35)
* Restricted LJUSB_isRecentKernel to linux only, return true on all other OSes

The code had a special case for macOS, but would probably not work correctly on BSD or Windows.  Instead, just use __linux__ preprocessor define to know if building for linux or not.

* Fixed a couple of incorrect comments

* Fixed a resource/memory leak in various error paths

In several places where there is an early return, libusb_free_device_list() was not being called.

* Fixed incorrect error handling in LJUSB_OpenAllDevices

The variable devh was initialized to NULL and then never changed. It was used in 2 calls to libusb_close(devh) but was always NULL when passed.  In fact, both calls to libusb_close() were using the wrong parameter, they should be libusb_close(handle).

Relatedly, the decrement of ljFoundCount in the error path just seems wrong, and would result in the function returning the wrong count of devices actually opened.

* Added new public API to open all devices of a given productId

This function doesn't require determining or guessing the maximim possible number of devices, instead it dynamically allocates an array big enough to hold them all.

It also allows you to limit yourself to specific productIds.

* Minor whitespace alignment fix